Creating new Star components to DRY out code
