+++ public/react/components/ItemDescription/ItemDescription.jsx
@@ -2,19 +2,18 @@ import React, { useState, useEffect, useContext, useRef } from "react";
 import { useParams, useNavigate } from "react-router";
 import { toast, Slide } from "react-toastify";
 import { AuthContext } from "../../AuthProvider";
-import { jwtDecode } from "jwt-decode";
 import UpdateItem from "../Forms/UpdateItemForm/UpdateItem";
 import DeleteItem from "../Forms/DeleteItemForm/DeleteItem";
 import apiURL from "../../api";
 import "react-toastify/dist/ReactToastify.css";
+import Stars from "../Stars/Stars";
 
 export default function ItemDescription() {
   const { id } = useParams();
   const navigate = useNavigate();
   const adminRef = useRef(null);
-  const { isLoggedIn, isAdmin } = useContext(AuthContext);
+  const { isLoggedIn, isAdmin, userId } = useContext(AuthContext);
   const [adminView, setAdminView] = useState(false);
-  const [userId, setUserId] = useState(null);
   const [singleItem, setSingleItem] = useState({
     id: 0,
     name: "",
@@ -30,41 +29,6 @@ export default function ItemDescription() {
       window.scrollTo({ top: document.body.scrollHeight, behavior: "smooth" });
     }, 200);
   }
-
-    function generateStars(num) {
-        const stars = [];
-        for (let i = 0; i < num; i++) {
-            stars.push(
-                <svg
-                key={i}
-                className="w-4 h-4 text-yellow-300"
-                aria-hidden="true"
-                xmlns="http://www.w3.org/2000/svg"
-                fill="currentColor"
-                viewBox="0 0 22 20"
-                >
-                <path d="M20.924 7.625a1.523 1.523 0 0 0-1.238-1.044l-5.051-.734-2.259-4.577a1.534 1.534 0 0 0-2.752 0L7.365 5.847l-5.051.734A1.535 1.535 0 0 0 1.463 9.2l3.656 3.563-.863 5.031a1.532 1.532 0 0 0 2.226 1.616L11 17.033l4.518 2.375a1.534 1.534 0 0 0 2.226-1.617l-.863-5.03L20.537 9.2a1.523 1.523 0 0 0 .387-1.575Z" />
-                </svg>
-            );
-        }
-        for (let i = num; i < 5; i++) {
-            stars.push(
-                <svg
-                key={i}
-                className="w-4 h-4 text-gray-200 dark:text-gray-600"
-                aria-hidden="true"
-                xmlns="http://www.w3.org/2000/svg"
-                fill="currentColor"
-                viewBox="0 0 22 20"
-                >
-                <path d="M20.924 7.625a1.523 1.523 0 0 0-1.238-1.044l-5.051-.734-2.259-4.577a1.534 1.534 0 0 0-2.752 0L7.365 5.847l-5.051.734A1.535 1.535 0 0 0 1.463 9.2l3.656 3.563-.863 5.031a1.532 1.532 0 0 0 2.226 1.616L11 17.033l4.518 2.375a1.534 1.534 0 0 0 2.226-1.617l-.863-5.03L20.537 9.2a1.523 1.523 0 0 0 .387-1.575Z" />
-                </svg>
-            );
-        }
-        return stars;
-    }
-
-    const num = Math.floor(Math.random() * 6);
 
   async function fetchSingleItem(id) {
     try {
@@ -113,14 +77,6 @@ export default function ItemDescription() {
 
   useEffect(() => {
     fetchSingleItem(id);
-    const token = localStorage.getItem("token");
-    if (token) {
-      const decoded = jwtDecode(token);
-      console.log(decoded);
-      if (decoded.id) {
-        setUserId(decoded.id);
-      }
-    }
   }, []);
 
   return (
@@ -145,9 +101,7 @@ export default function ItemDescription() {
         <div className="p-4 flex flex-col justify-around flex-wrap w-1/2">
           <h2 className="text-xl font-semibold text-gray-800">
             {singleItem.name}
-            <div className="flex items-center space-x-1 rtl:space-x-reverse bg-white">
-              {generateStars(num)}
-            </div>
+            <Stars />
           </h2>
           <p className="text-gray-600 mt-2">{singleItem.description}</p>
           <div className="mt-4 flex justify-between items-center">
